DoCheckerBack/requirements.txt
Django~=4.0.3
psycopg2
django-environ>=0.4.5
django-extensions==3.1.3
django-stubs==1.8.0
djangorestframework
django-elasticsearch-dsl
django-elasticsearch-dsl-drf
django-rest-swagger
pyyaml
gunicorn==20.1.0
django-cors-headers
psycopg2
